+++ packages/aws/src/schemas/s3-files.dimensions.ts
@@ -1,4 +1,4 @@
-import { getDimensionsSchema } from 'sanity-plugin-external-dam'
+import { getDimensionsSchema } from 'sanity-plugin-external-files'
 import config from '../config'
 
 export default getDimensionsSchema(config)