+++ packages/aws/src/schemas/s3-files.media.ts
@@ -1,10 +1,10 @@
 import { defineField } from 'sanity'
-import { Rule } from '@sanity/types'
 import Input from '../components/Input'
 
 export default defineField({
-  name: 'digital-ocean-files.media',
-  title: 'Digital Ocean media',
+  // @TODO: how to handle schema/id changes?
+  name: 's3-files.media',
+  title: 'S3 media',
   type: 'object',
   components: {
     input: Input as any,
@@ -14,8 +14,8 @@ export default defineField({
       name: 'asset',
       title: 'Asset',
       type: 'reference',
-      to: [{ type: 'digital-ocean-files.storedFile' }],
-      validation: (Rule: Rule) => Rule.required(),
+      // @TODO: how to handle schema/id changes?
+      to: [{ type: 's3-files.storedFile' }],
     },
   ],
 })