+++ packages/aws/src/tool.ts
@@ -2,11 +2,9 @@ import { ToolIcon } from 'sanity-plugin-external-files'
 import config from './config'
 import Tool from './components/Tool'
 
-const DigitalOceanFilesTool = {
-  name: 'digital-ocean-files',
+export default {
+  name: 's3-files',
   title: config.toolTitle,
   component: Tool,
   icon: ToolIcon,
 }
-
-export default DigitalOceanFilesTool