+++ packages/core/src/index.ts
@@ -1,4 +1,4 @@
-import CreateInput from './components/CreateInput'
+import createInput from './components/CreateInput'
 import StudioTool from './components/StudioTool'
 import ToolIcon from './components/ToolIcon'
 import getStoredFileSchema from './schemas/getStoredFileSchema'
@@ -6,7 +6,7 @@ import getDimensionsSchema from './schemas/getDimensionsSchema'
 import getCustomDataSchema from './schemas/getCustomDataSchema'
 
 export {
-  CreateInput,
+  createInput,
   StudioTool,
   ToolIcon,
   getStoredFileSchema,