+++ packages/firebase/src/index.ts
@@ -1,10 +1,21 @@
-import { ToolIcon } from 'sanity-plugin-external-dam'
-import config from './config'
-import Tool from './components/Tool'
+import { definePlugin } from 'sanity'
+import firebaseFilesCustomData from './schemas/firebase-files.custom-data'
+import firebaseFilesDimensions from './schemas/firebase-files.dimensions'
+import firebaseFilesStoredFile from './schemas/firebase-files.storedFile'
+import firebaseFilesMedia from './schemas/firebase-files.media'
+import FirebaseFilesTool from './tool'
 
-export default {
-  name: 'firebase-dam',
-  title: config.toolTitle,
-  component: Tool,
-  icon: ToolIcon,
-}
+export const firebaseFiles = definePlugin(() => {
+  return {
+    name: 'firebase-files',
+    schema: {
+      types: [
+        firebaseFilesCustomData,
+        firebaseFilesDimensions,
+        firebaseFilesStoredFile,
+        firebaseFilesMedia,
+      ],
+    },
+    tools: [FirebaseFilesTool],
+  }
+})