+++ packages/firebase/src/schemas/firebase-files.media.ts
@@ -1,21 +1,21 @@
 import { defineField } from 'sanity'
-import { Rule } from '@sanity/types'
 import Input from '../components/Input'
 
 export default defineField({
-  name: 'digital-ocean-files.media',
-  title: 'Digital Ocean media',
+  // @TODO: how to handle schema/id changes?
+  name: 'firebase-files.media',
+  title: 'Firebase media',
   type: 'object',
   components: {
-    input: Input as any,
+    input: Input as any
   },
   fields: [
     {
       name: 'asset',
       title: 'Asset',
       type: 'reference',
-      to: [{ type: 'digital-ocean-files.storedFile' }],
-      validation: (Rule: Rule) => Rule.required(),
+      // @TODO: how to handle schema/id changes?
+      to: [{ type: 'firebase-files.storedFile' }],
     },
   ],
 })